Extract ResourceScalaCompat so the other parts can be shared by Scala 2 and 3
